+++ pkg/hmenum/rega_script.go
@@ -25,6 +25,7 @@ const (
 	RegaScriptGetSystemUpdateInfo           RegaScript = "get_system_update_info"
 	RegaScriptGetSystemVariableDescriptions RegaScript = "get_system_variable_descriptions"
 	RegaScriptSetProgramState               RegaScript = "set_program_state"
+	RegaScriptExecuteProgramConditional     RegaScript = "execute_program_conditional"
 	RegaScriptSetSystemVariable             RegaScript = "set_system_variable"
 	// Lifecycle scripts. Names match
 	// Reference where one exists
@@ -67,6 +68,7 @@ var AllRegaScripts = []RegaScript{
 	RegaScriptGetSystemUpdateInfo,
 	RegaScriptGetSystemVariableDescriptions,
 	RegaScriptSetProgramState,
+	RegaScriptExecuteProgramConditional,
 	RegaScriptSetSystemVariable,
 	RegaScriptCreateSystemVariable,
 	RegaScriptUpdateSystemVariable,